refactor: standardize file naming convention
